"Updated README file with same content"
Explanation: The commit message should simply describe the changes made in the code diff. In this case, the diff shows that there is no difference in the content of the README file, but only a change in the commit history (indicated by the index values). Therefore, the commit message should reflect this by stating that the README file was updated with the same content.